requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.3.1            # via django
attrs==20.3.0             # via pytest
certifi==2020.12.5        # via requests
chardet==3.0.4            # via requests
coreapi==2.3.3            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
django-cors-headers==3.6.0  # via -r requirements.in
django-environ==0.4.5     # via -r requirements.in
django-phonenumber-field[phonenumbers]==5.0.0  # via -r requirements.in
django-rest-swagger==2.2.0  # via -r requirements.in
django==3.1.4             # via -r requirements.in, django-cors-headers, django-phonenumber-field, djangorestframework
djangorestframework-jwt==1.11.0  # via -r requirements.in
djangorestframework==3.12.2  # via -r requirements.in, django-rest-swagger
idna==2.10                # via requests
iniconfig==1.1.1          # via pytest
itypes==1.2.0             # via coreapi
jinja2==2.11.2            # via coreschema
markupsafe==1.1.1         # via jinja2
openapi-codec==1.3.2      # via django-rest-swagger
packaging==20.8           # via pytest
phonenumbers==8.12.15     # via django-phonenumber-field
pillow==8.0.1             # via -r requirements.in
pluggy==0.13.1            # via pytest
psycopg2-binary==2.8.6    # via -r requirements.in
py==1.10.0                # via pytest
pyjwt==1.7.1              # via djangorestframework-jwt
pyparsing==2.4.7          # via packaging
pytest-django==4.1.0      # via -r requirements.in
pytest==6.2.1             # via pytest-django
pytz==2020.4              # via django
requests==2.25.0          # via coreapi
simplejson==3.17.2        # via django-rest-swagger
sqlparse==0.4.1           # via django
toml==0.10.2              # via pytest
uritemplate==3.0.1        # via coreapi
urllib3==1.26.2           # via requests
